+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ### 2.0.2
 
 - Bump FBSDK to 4.27.1
+- Clean 
 
 ### 2.0.1